Add files via upload
I'm uploading all three versions.  The original somehow has the latest date.  I may have continued to work on the original and the others were just duplicates at that moment, which is fine for what it was used for.  But, you need the final version which is either the original or 3.